requirements.txt
pyTelegramBotAPI
google-generativeai
telethon
requests
beautifulsoup4
feedparser
python-dotenv
